backend/pyproject.toml
[tool.poetry]
name = "talent-lens-backend"
version = "0.1.0"
description = "Backend service for TalentLens platform"
authors = ["tkhongsap <khongsap@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.1"
uvicorn = "^0.24.0"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic = "*"
python-dotenv = "*"
motor = "^3.3.2"
python-magic-bin = "^0.4.14"
numpy = "^1.24.0"
spacy = "^3.7.2"
llama-parse = "*"
openai = "*"
colorama = "*"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
httpx = "^0.25.2"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi-line-output = 3 